requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
boto3==1.27.0
botocore==1.30.0
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
DateTime==5.1
discord-webhook==1.1.0
discord.py==2.2.0
idna==3.3
jmespath==1.0.1
multidict==6.0.2
packaging==23.0
pip-check==2.8.1
pycares==4.2.1
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
terminaltables==3.1.10
typing_extensions==4.3.0
urllib3==1.26.16
yarl==1.7.2
zope.interface==6.0
